requirements.txt
_tflow_select=2.3.0=mkl
import tensorflow as tf
absl-amly=0.5.0=py36_0
astor=0.7.1=py36_0
blas=1.0=mkl
ca-certificates=2019.3.9=hecc5488_0
certifi=2019.3.9=py36_0
cloudpickle=0.6.1=py36_0
cycler=0.10.0=py36_0
dask-core=0.19.4=py36_0
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py36_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gast=0.2.0=py36_0
gdcm=2.8.9=py36h71b2a6d_0
glib=2.56.2=hd408876_0
grpcio=1.12.1=py36hdbcaa40_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py36h989c5e5_3
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
imageio=2.4.1=py36_0
intel-openmp=2019.0=118
jpeg=9b=h024ee3a_2
json-c=0.13.1=h14c3975_1001
keras=2.2.4=0
keras-applications=1.0.6=py36_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.0.5=py36_0
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libjpeg-turbo=2.0.2=h14c3975_0
libpng=1.6.35=hbc83047_0
libprotobuf=3.6.0=hdbcaa40_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markdown=3.0.1=py36_0
matplotlib=3.0.0=py36h5429711_0
mkl=2019.0=118
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
ncurses=6.1=hf484d3e_0
networkx=2.2=py36_1
numpy=1.15.2=py36h1d66e8a_1
numpy-base=1.15.2=py36h81de0dd_1
olefile=0.46=py36_0
openjpeg=2.3.1=h58a6597_0
openssl=1.1.1b=h14c3975_1
pandas=0.23.4=py36h04863e7_0
pcre=8.42=h439df22_0
pillow>=8.3.2
pip>=21.1
protobuf=3.6.0=py36hf484d3e_0
pyparsing=2.2.2=py36_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.8=h0371630_0
python-dateutil=2.7.3=py36_0
pytz=2018.5=py36_0
pywavelets=1.0.1=py36hdd07704_0
pyyaml=5.4
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scikit-image=0.14.0=py36hf484d3e_1
scikit-learn=0.20.0=py36h4989274_1
scipy=1.1.0=py36hfa4b5c9_1
setuptools=40.4.3=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.11.0=py36_1
sqlite=3.26.0=h7b6447c_0
tensorboard=>=1.12.1
tensorflow>=1.12.1
tensorflow-base>=1.12.1
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
toolz=0.9.0=py36_0
tornado=5.1.1=py36h7b6447c_0
werkzeug=0.15.3
wheel=0.32.1=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=ha838bed_2
